requirements/production.txt
amqp==2.3.2
anyjson==0.3.3
apscheduler==3.6.0
kafka-python==1.4.6
web3==4.9.2
beautifulsoup4==4.6.1
billiard>=3.3.0.23,<3.4
boto3==1.7.58
botocore==1.10.58
bs4==0.0.1
celery==3.1.20
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.6
django-celery==3.2.2
django-cors-headers==2.3.0
django-debug-toolbar==1.9.1
django-enumfields==0.10.1
django-environ==0.4.5
django-extensions==2.1.0
django-filter==2.0.0
django-redis==4.9.0
django-redis-cache==1.7.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
docutils==0.14
drf-enum-field==0.9.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
kombu>=3.0.37,<3.1
MarkupSafe==1.0
openapi-codec==1.3.2
passlib==1.7.1
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pybase62==0.3.3
pycryptodome==3.8.1
python-dateutil==2.7.3
python-magic==0.4.15
pytz==2018.5
raven==6.9.0
redis==2.10.6
requests==2.19.1
s3transfer==0.1.13
simplejson==3.16.0
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.23
uWSGI==2.0.17.1
vine==1.1.4
social-auth-app-django==3.1.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==3.2.0
social-auth-storage-mongoengine==1.0.1
elastic-apm==5.3.1
django-nine==0.2.2
django-elasticsearch-dsl==7.1.0
django-elasticsearch-dsl-drf==0.20.3
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
